provider+tabs/bmm: move mining logic to provider
we don't want to lose all history when routing away from the bmm-tab
