+++ lib/pages/tabs/testchain/mainchain/bmm_tab_page.dart
@@ -1,9 +1,7 @@
 import 'dart:async';
 import 'dart:convert';
-import 'dart:math';
 
 import 'package:auto_route/auto_route.dart';
-import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
 import 'package:sail_ui/sail_ui.dart';
@@ -11,8 +9,8 @@ import 'package:sail_ui/widgets/core/sail_text.dart';
 import 'package:sidesail/bitcoin.dart';
 import 'package:sidesail/logger.dart';
 import 'package:sidesail/pages/tabs/dashboard_tab_page.dart';
+import 'package:sidesail/providers/bmm_provider.dart';
 import 'package:sidesail/rpc/models/bmm_result.dart';
-import 'package:sidesail/rpc/rpc_testchain.dart';
 import 'package:sidesail/widgets/containers/tabs/dashboard_tab_widgets.dart';
 import 'package:stacked/stacked.dart';
 
@@ -133,7 +131,7 @@ class BlindMergedMiningTabPage extends StatelessWidget {
                                     itemBuilder: (context, index) => BMMAttemptView(
                                       key: ValueKey<String>(viewModel.bmmAttempts[index].result.txid),
                                       bmmAttempt: viewModel.bmmAttempts[index],
-                                      onPressed: () => viewModel.toggleRunning(false),
+                                      onPressed: () => {},
                                     ),
                                   ),
                                 ],
@@ -152,7 +150,7 @@ class BlindMergedMiningTabPage extends StatelessWidget {
 }
 
 class BlindMergedMiningTabPageViewModel extends BaseViewModel {
-  TestchainRPC get _rpc => GetIt.I.get<TestchainRPC>();
+  BMMProvider get _bmmProvider => GetIt.I.get<BMMProvider>();
 
   final bidController = TextEditingController();
   final refreshController = TextEditingController();
@@ -163,11 +161,12 @@ class BlindMergedMiningTabPageViewModel extends BaseViewModel {
   BlindMergedMiningTabPageViewModel() {
     bidController.text = '1000';
     refreshController.text = '3';
+
+    _bmmProvider.addListener(notifyListeners);
   }
 
   bool running = false;
-
-  List<BmmAttempt> bmmAttempts = [];
+  List<BmmAttempt> get bmmAttempts => _bmmProvider.bmmAttempts;
 
   Timer? refreshTimer;
 
@@ -178,7 +177,9 @@ class BlindMergedMiningTabPageViewModel extends BaseViewModel {
     if (running) {
       log.i('starting refresh timer with interval $refreshSeconds');
 
-      refreshTimer = Timer.periodic(Duration(seconds: refreshSeconds), _handleBmmTick);
+      refreshTimer = Timer.periodic(Duration(seconds: refreshSeconds), (_) {
+        _bmmProvider.handleBmmTick(bidAmountSats);
+      });
     } else {
       log.i('stopping refresh timer');
 
@@ -187,96 +188,10 @@ class BlindMergedMiningTabPageViewModel extends BaseViewModel {
     }
   }
 
-  // Based off of RefreshBMM in testchain
-  // https://github.com/LayerTwo-Labs/testchain/blob/c44e3a737d1780a1a07135657ac0fd7686251933/src/qt/sidechainpage.cpp#L767
-  void _handleBmmTick(Timer _) async {
-    final res = await _rpc.refreshBMM(bidAmountSats);
-    final currentBlockCount = await _rpc.sideBlockCount();
-
-    if (res.error != null) {
-      log.d('was not able to advance BMM: ${res.error}');
-    }
-
-    var attempt = BmmAttempt(
-      result: res,
-      bidSatoshis: bidAmountSats,
-
-      // BMM attempts are always for the /next/ block height
-      sidechainBlockHeight: currentBlockCount + 1,
-
-      // TODO: fetch mainchain block by TXID to get a correct number
-      mainchainBlockHeight: await _rpc.mainBlockCount(),
-    );
-
-    // Add the attempt!
-    if (res.bmmBlockCreated != null) {
-      log.d('adding attempt: ${res.bmmBlockCreated}');
-      _addAttempt(attempt);
-    }
-
-    // We were able to advance the chain, and must update its descendants.
-    // (Or something like that... Not completely sure, tbh)
-    if (res.bmmBlockSubmitted != null) {
-      log.d('updating connected: ${res.bmmBlockSubmitted}');
-      _updateConnected(res);
-    }
-
-    notifyListeners();
-  }
-
-  // Ported from SidechainBMMTableModel::AddAttempt in the testchain codebase
-  void _addAttempt(BmmAttempt attempt) {
-    bmmAttempts.insert(
-      0,
-      attempt,
-    );
-
-    // Skip if there are less than 2 BMM requests
-    if (bmmAttempts.length < 2) {
-      log.d('less than twp BMM requests, skipping');
-      return;
-    }
-
-    // Skip the first (newest) entry
-    const start = 1;
-    // When we add a new attempt, set the last 6 BMM requests to failed if
-    // they weren't connected - they have expired.
-    const attemptsToUpdate = 6;
-    bmmAttempts.slice(start, min(start + attemptsToUpdate, bmmAttempts.length)).forEach((attempt) {
-      if (attempt.connected) {
-        return;
-      }
-
-      log.d('marking attempt as failed: ${attempt.result.txid}');
-
-      // If the attempt wasn't connected, it per def failed.
-      attempt.failed = true;
-    });
-  }
-
-  // Ported from SidechainBMMTableModel::UpdateForConnected in the testchain codebase
-  void _updateConnected(BmmResult current) {
-    //  auto it = std::find_if(model.begin(), model.end(),
-    //     [ hashMerkleRoot ](BMMTableObject a) { return a.hashMerkleRoot == hashMerkleRoot; });
-    final toUpdate =
-        bmmAttempts.firstWhereOrNull((prev) => prev.result.bmmBlockCreated == current.bmmBlockSubmittedBlind);
-
-    if (toUpdate == null) {
-      log.w('was not able to find bmmBlockSubmittedBlind: ${current.bmmBlockSubmittedBlind}');
-      return;
-    }
-
-    log.d('updating connected block: ${toUpdate.result.txid}');
-
-    //  it->fFailed = false;
-    // it->fConnected = true;
-    toUpdate.failed = false;
-    toUpdate.connected = true;
-  }
-
   @override
   void dispose() {
     refreshTimer?.cancel();
+    _bmmProvider.removeListener(notifyListeners);
     super.dispose();
   }
 }
@@ -381,7 +296,7 @@ class _BMMAttemptViewState extends State<BMMAttemptView> {
 
   String extractTXTitle(BmmAttempt attempt) {
     String title =
-        '${(satoshiToBTC(attempt.bidSatoshis)).toStringAsFixed(8)} BTC bid containing ${attempt.result.ntxn} transaction(s) with ${attempt.result.nfees.toStringAsFixed(8)} SBTC total fees';
+        '${(satoshiToBTC(attempt.bidSatoshis)).toStringAsFixed(8)} BTC bid containing ${attempt.result.ntxn} transaction(s) with ${satoshiToBTC(attempt.result.nfees).toStringAsFixed(8)} SBTC total fees';
     return title;
   }
 }
